requirements.txt
openai==1.50.2
streamlit==1.39.0
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.3